requirements.txt
matplotlib==3.10.1
numpy==2.2.5
pandas==2.2.3
plotly==6.0.1
PuLP==3.0.2
python-dotenv==1.1.0
scipy==1.15.2
setuptools==72.1.0
SQLAlchemy==2.0.40
streamlit==1.44.1


